requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.11.0
openai==1.54.5
python-dotenv==1.0.0
numpy==1.26.4
scikit-learn==1.5.2
anthropic==0.39.0
gunicorn==21.2.0